+++ Makefile
@@ -9,7 +9,7 @@ TARGET = read
 TEST_TARGET = test_exec
 
 # Source file
-SRC = src/read.cpp src/network.cpp src/layer.cpp src/labeledvector.cpp
+SRC = src/read.cpp src/network.cpp src/layer.cpp src/labeledvector.cpp src/read_helpers.cpp
 
 # Test files
 TEST_SRC = tests/test_main.cpp tests/test_layer.cpp tests/catch_amalgamated.cpp src/layer.cpp
