+++ tests/test_main.cpp
@@ -1,2 +1,58 @@
 #define CATCH_CONFIG_MAIN
 #include "catch_amalgamated.hpp"
+// data = [
+//     ("Fantastic spot for an even or a quite cocktail. They were swell to host the Yelp crew with a great drink menu and super attentive staff.I'd certainly recommend anything with the purred fruit in it (apple, any of them really)!", 1),
+//     ("Love, love, love the calamari. It's so good and spicy. There is an endless list of martinis and one is better than the next. There is a first and second floor, if you prefer more of a bar or lounge atmosphere. It's not a place I like for dinner, but for drinks and apps it's wonderful.", 1),
+//     ("Love this place. Stiff martinis and cocktails, cheap drinks, good service, nice atmosphere to chill in the upstairs lounge and hang up with your friends. Classy crowd and much more mature, older, and professional crowd. There aren't any of those college frat boys or belligerent drunks there, which is so great. Very nice place to wind down after a long day.", 1),
+//     ("It's everything a great cocktail bar should be. Great location, ambiance, cocktails and chill environment.**Disclaimer: 5 stars doesn't have to mean it's New York City good. It is a simple place and it's not flashy by any means. I'd take a girl for the beginning of a date or on the opposite end of the spectrum, talking business with colleagues.", 1),
+//     ("I came here before a pirates game, so it was around 5:30ish or so in the evening, we had just gotten lucky that they had a table open before all there reservations for the night started to come in. I ordered the chicken jerk sandwich with a side salad, absolutely loved it! The chicken had the perfect amount of spice to it which was off set with sweet with the pineapple chutney it had on it, sooo good! The salad was even great because they used real green spring mix instead of iceburg lettuce. I also loved that they put feta cheese on top, its one of my favorite cheeses for salad topping! Service was great here and even though I didn't have a drink here that night I have a previous time and they are yummy! Love this place and can't wait for my next visit!", 1),
+//     ("My wife and I caught this show at the Golden Nugget and had heard good things about him in the media, so we were pretty excited to see him. 3 minutes into the show, we were staring at each other in disbelief. Absolutely the corniest sense of humor I've ever had the misfortune of seeing on a stage. His impressions are often dated and usually inaccurate. If he hadn't announced some of them prior to performing them, I would have been at a loss for who he was mimicking. A half hour into the show, we were waiting for him to turn his back on the crowd for a moment so we could run out of the theater and put an end to our suffering. I do acknowledge that some people in the audience seemed to be enjoying themselves, but I have no idea who his audience is. Maybe people who thought \"Mama's Family\" was the funniest show ever. I've heard people compare him to Danny Gans, but as far as I'm concerned, there's no comparison. Gans was much more entertaining, his impressions were much more spot-on, and although I didn't find him the funniest guy in the world, he was Richard Pryor compared to Mr. Brown. I would certainly recommend a trip to Frank Caliendo's show if you're looking for an impressionist. If you find yourself at the Golden Nugget, go stare at their aquarium rather than watch this show...it's more entertaining and they have a pufferfish that does a better Nicholson than Brown does.", 0),
+//     ("This was the dumbest show we have ever seen. We never laughed once and after 15 minutes we realized we would have 2 more hours of the same boring stupidity. We're not sure what some of the other reviewers here saw, but there were many people around us who were bored too and never clapped. Music was extremely loud and theater seating is on level ground and you are looking over the heads in front of you. Save your money and don't go! He used to be on the strip but has been demoted to Golden Nugget downtown. We went to see Carrot Top at the Luxor on the strip. OMG!!! Now Carrot Top is a show. We laughed our heads off. Put your money on Carrot Top's show and go to that!!", 0),
+//     ("My girlfriend and I went to this show. It was absolutely terrible!!!Not funny at all. We never laughed once and after 5 minutes wanted to get the hell out of there!!!We actually felt sorry for him because it was so pitiful.And it was ridiculously loud!Not sure what the others are posting 4 star reviews forLuckily we got our tickets for free. If I had paid good money for them I would be furious.Save your money and do not go to this show.", 0),
+//     ("The restrooms looked like after an bombardment, an improvised shack amid buckets of olives and spices. Highly questionable hygiene for a food store!Disgusting. I will not go back until they clean up the place. It's just filthy!!!", 0),
+//     ("Have been going to the Pennsylvania Macaroni Company for years. Love their fresh pasta and great selection of cheeses. Always great service, however today was the exception. Rudest woman at the cheese counter. Then went to get the fresh pasta and the woman behind the counter was downright nasty. She was covering the pasta at 4:00, they close at 4:30 and I asked her if it was to late to get pasta. She rolled her eyes at me and said uh yeah-we are putting it away right now. A simple yes or no would have been nice. I said ok thanks anyway. Then she must have reconsidered and said well did you want something In a very rude voice. Needless to say I walked away. Wow! Pennsylvania Macaroni Company needs to work on customer service. Obviously she hates her job, then get another one! If I didn't drive so far to get there I wouldn't have written this review. Every January I buy 3 months worth of cheese for my family members. I won't be doing that anymore.", 0)
+// ]
+
+// --- Review 1 (Label: 1) ---
+// fantastic           : 0.2359
+// spot                : 0.2006
+// quite               : 0.2359
+// cocktail            : 0.2006
+// swell               : 0.2359
+// host                : 0.2359
+// yelp                : 0.2359
+// crew                : 0.2359
+// great               : 0.1401
+// drink               : 0.2006
+// menu                : 0.2359
+// super               : 0.2359
+// attentive           : 0.2359
+// staff               : 0.2359
+// certainly           : 0.2006
+// recommend           : 0.2006
+// purred              : 0.2359
+// fruit               : 0.2359
+// apple               : 0.2359
+// really              : 0.2359
+
+// --- Review 2 (Label: 1) ---
+// love                : 0.4477
+// calamari            : 0.2257
+// good                : 0.1212
+// spicy               : 0.2257
+// endless             : 0.2257
+// list                : 0.2257
+// martinis            : 0.1918
+// better              : 0.1918
+// second              : 0.2257
+// floor               : 0.2257
+// prefer              : 0.2257
+// bar                 : 0.1918
+// lounge              : 0.1918
+// atmosphere          : 0.1918
+// place               : 0.1340
+// like                : 0.1918
+// dinner              : 0.2257
+// drinks              : 0.1918
+// apps                : 0.2257
+// wonderful           : 0.2257
