+++ tests/test_read_helpers.cpp
@@ -1,0 +1,2 @@
+#include "catch_amalgamated.hpp"
+#include "../src/read_helpers.h"
